tests/wpt/web-platform-tests/html/webappapis/scripting/events/event-handler-attributes-body-window.html

<!DOCTYPE html>
<title>HTMLBodyElement event handlers</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/WebIDLParser.js"></script>
<script src="resources/event-handler-body.js"></script>
<div id="log"></div>
<body>
<script>
setup({ explicit_done: true });

handlersListPromise.then(({ shadowedHandlers, notShadowedHandlers }) => {
  const createdBody = document.createElement("body");
  for (const [description, body, altBody] of [
    ["document.body", document.body, createdBody],
    ['document.createElement("body")', createdBody, document.body]
  ]) {
    const f = () => 0;

    shadowedHandlers.forEach(function(handler) {
      test(function() {
        body['on' + handler] = f;
        assert_equals(window['on' + handler], f, "window should reflect");
        assert_equals(altBody['on' + handler], f, "alternative body should reflect");
      }, `shadowed ${handler} (${description})`);
    });
    notShadowedHandlers.forEach(function(handler) {
      test(function() {
        body['on' + handler] = f;
        assert_equals(window['on' + handler], null, "window should reflect");
        assert_equals(altBody['on' + handler], null, "alternative body should reflect");
      }, `not shadowed ${handler} (${description})`);
    });

    [...shadowedHandlers, ...notShadowedHandlers].forEach(function(handler) {
      body['on' + handler] = null;
    });

    shadowedHandlers.forEach(function(handler) {
      test(function() {
        assert_equals(body['on' + handler], null, "body should reflect changes to itself");
        assert_equals(window['on' + handler], null, "window should reflect");
        assert_equals(altBody['on' + handler], null, "alternative body should reflect");
      }, `shadowed ${handler} removal (${description})`);
    });
  }

  done();
});
</script>
